refactored distribution to its own results controller
