+++ config/routes.rb
@@ -58,8 +58,10 @@ Rails.application.routes.draw do
       scope module: :jwt_api do
         resources :one_time_response, only: [:index, :show], param: :otr
         resources :questionnaire, only: [:show, :create], param: :key do
-          member do
-            get :distribution
+          resources :results, only: [] do
+            collection do
+              get :distribution
+            end
           end
         end
         resources :response, only: [:show, :index, :create], param: :uuid do
